Improve parser for kfc
